requirements.txt
absl-py==2.1.0
anyio==4.3.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
attrs==23.2.0
Babel==2.14.0
beautifulsoup4==4.12.3
bleach==6.1.0
bleak==0.22.2
blinker==1.8.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.1
contourpy==1.2.0
cycler==0.12.1
debugpy==1.8.1
decorator==5.1.1
defusedxml==0.7.1
dotmap==1.3.30
executing==2.0.1
fastjsonschema==2.19.1
Flask==3.0.3
flatbuffers==24.3.25
fonttools==4.48.1
fqdn==1.5.1
future==1.0.0
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
idna==3.6
ipykernel==6.29.3
ipython==8.22.1
iso8601==2.1.0
isoduration==20.11.0
itsdangerous==2.2.0
jax==0.4.29
jaxlib==0.4.29
jedi==0.19.1
Jinja2==3.1.3
json5==0.9.19
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter-events==0.9.0
jupyter-lsp==2.2.3
jupyter_client==8.6.0
jupyter_core==5.7.1
jupyter_server==2.12.5
jupyter_server_terminals==0.5.2
jupyterlab==4.1.2
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.3
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.8.2
matplotlib-inline==0.1.6
mediapipe==0.10.14
meshtastic==2.3.11
mistune==3.0.2
ml-dtypes==0.4.0
nbclient==0.9.0
nbconvert==7.16.1
nbformat==5.9.2
nest-asyncio==1.6.0
notebook==7.1.1
notebook_shim==0.2.4
numpy==1.26.4
opencv-contrib-python==4.9.0.80
opencv-python==4.9.0.80
opt-einsum==3.3.0
overrides==7.7.0
packaging==23.2
pandas==2.2.1
pandocfilters==1.5.1
parso==0.8.3
pexpect==4.9.0
pillow==10.2.0
platformdirs==4.2.0
prometheus_client==0.20.0
prompt-toolkit==3.0.43
protobuf==5.27.1
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
pygame==2.5.2
Pygments==2.17.2
pyparsing==3.1.1
Pypubsub==4.0.3
PyQRCode==1.2.1
pyserial==3.5
python-dateutil==2.8.2
python-json-logger==2.0.7
pytz==2024.1
pywin32==306
pywinpty==2.0.13
PyYAML==6.0.1
pyzmq==25.1.2
referencing==0.33.0
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.18.0
scipy==1.13.1
Send2Trash==1.8.2
six==1.16.0
sniffio==1.3.1
sounddevice==0.4.7
soupsieve==2.5
stack-data==0.6.3
tabulate==0.9.0
terminado==0.18.0
timeago==1.0.16
tinycss2==1.2.1
tornado==6.4
traitlets==5.14.1
types-python-dateutil==2.8.19.20240106
tzdata==2024.1
uri-template==1.3.0
urllib3==2.2.1
wcwidth==0.2.13
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
Werkzeug==3.0.3
winrt-runtime==2.0.1
winrt-Windows.Devices.Bluetooth==2.0.1
winrt-Windows.Devices.Bluetooth.Advertisement==2.0.1
winrt-Windows.Devices.Bluetooth.GenericAttributeProfile==2.0.1
winrt-Windows.Devices.Enumeration==2.0.1
winrt-Windows.Foundation==2.0.1
winrt-Windows.Foundation.Collections==2.0.1
winrt-Windows.Storage.Streams==2.0.1
